assignments/first-assignment/ronyz/requirements.txt
mangum
fastapi
requests
bs4
python-dateutil
datetime
boto3
